feat(firebase): Add customer approval events and customers-only store config
Stores that must approve customers before selling can now subscribe to
`customers-staffSignatureSet` and `customers-enabledSet` API events
(mapped with `modified_fields`, same pattern as `products-priceSet`)
and declare `passport.newCustomer` / `checkout.*` gates in config.

- `parseEventName` moved to its own side-effect-free module with unit tests
- new `CheckoutGatesConfig` and `PassportConfig` types on `config.get()`
